Update 1 : Better looking plots overall
Have added two seperate HTML files for each different generated plots, have changes the background colors to black and have made an okay radius scale for plotly markers in each plots. The hover-templates have been changed to display the names of the pointed stars. A function was created to extract a proper name for the given stars of the catalogue in accordance to the database. The radii displayed are quite weird, especially the ones displayed in the HR diagram where we can see small stars in the giants and supergiants sections of the diagram, hinting at a possible mishaps or bug. I will need to look into it more.